Title: 십진수와 이진수, Time: 113ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240317/십진수와 이진수/decimal-and-binary-number.java
@@ -11,7 +11,7 @@ public class Main {
 
         num *= 17;
 
-        Stack<Integer> s = new Stack<>();
+        Stack<Long> s = new Stack<>();
         while(num > 0){
             s.add(num % 2);
             num /= 2;
